app/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
geoalchemy2
asyncpg
alembic
pydantic
typing
uuid
redis
databases
fastapi_limiter
python-dotenv